Remove old minio IP workaround
This workaround was needed back when we ran Django in a container
in development, too. Nowadays we run Django in a normal venv,
and only use containers for Postgres and Minio.

Back then, the problem was that both the Django container and your browser
would need to connect to Minio. From Django, only "tq-data" and "172.x.x.x"
would be valid ways (but not localhost). From your browser, localhost and
172.x.x.x would work, but not tq-data (the container name of minio).

Now that both Django and your browser run directly on your laptop, we can
simply use localhost. Or rather to avoid issues, 127.0.0.1.